Move cache management to a dedicated class and in the setting module, for clarity.
