+++ libraries/matrix/api/src/main/kotlin/io/element/android/libraries/matrix/api/auth/MatrixAuthenticationService.kt
@@ -30,13 +30,6 @@ interface MatrixAuthenticationService {
     suspend fun login(username: String, password: String): Result<SessionId>
 
     /*
-     * Cache index
-     */
-
-    fun cacheIndex(): Flow<Int>
-    fun incrementCacheIndex()
-
-    /*
      * OIDC part.
      */
